Add space and time complexity analysis to section-05-02 and create section-05-03
This commit includes updates to two JavaScript files in the faang/section-05 directory. In faang/section-05/section-05-02.js, extraneous lines were removed and a detailed analysis of the space and time complexity of the 'lengthOfLongestSubstring' function was added. Additionally, a new script, section-05-03.js, was created which appears to demonstrate a "sliding window" technique in the same 'lengthOfLongestSubstring' function. Both scripts now include explanations of their space and time complexity.