Add function to reverse linked list between positions
This commit introduces a new JavaScript function called `reverseBetween` in the `faang` directory. This function is designed to reverse a linked list only in between the positions `m` and `n` (both inclusive). A new markdown file is also added that contains a detailed breakdown and explanation of how the function works.